Extract strings.c from misc2.c
